users_management_service/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.3.3
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
drf-spectacular==0.27.0
h11==0.14.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mysql-connector-python==8.2.0
mysqlclient==2.2.0
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.21.12
pycparser==2.21
pyodbc==5.0.1
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-mock==1.11.0
responses==0.24.1
rpds-py==0.13.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0